refactor: contextapi적용
